Enemy Movement (Initial Completion) - To Goals and to Player
Enemy will now move to Player Character (Primary Target) ignoring its goals if player too nearby. If enemy goes too far away from its goals, itll go back to it. If enemy character is unable to reach certain goals (invalid or unreachable) then it removes that goal and creates a new goal location for it to go to.

Next in line to come:
-Transforming Platforms.
- Doors/Gates.
- NPCs.
- Weapons + Magic.
- Enemy/NPC Attacks.
- Player Attacks.
- World Designing? (earlier?).
- Immovable Enemies (Statues)